Deploy manifest (#20)
* add delete_manifest_stage and deploy_manifest_stage

deploy_manifest_stage was tricky because it allows
arbitrary manifest in the definition. It was expected
to be the yaml manifest, but spinnaker likes to convert
it to json. So had to do some yaml <-> json

* fix test wording

* remove printlns

* fix deploy manifest source serialize to be text

* add delete manifest provider stage

* add deploy/delete manifest to provider

* fix deploy manifest resource diff

* spinnaker error timestamp to be string

* update spinnaker error to better handle timestamp

timestamp could be any data format, so don't fail to unmarshal
also let `~` be in the cert/key path